refactor: clean coding pages folder structure
