+++ app/dashboard/[id]/_components/ItemStatus/types.ts
@@ -1,0 +1,3 @@
+export interface IStatusForm {
+  status: string;
+}
