+++ app/dashboard/[id]/_components/RealtimeItem/types.ts
@@ -1,0 +1,7 @@
+// TYPES
+import { IQueryItem, ICommentItem } from "@/types/db";
+
+export interface RealtimeItemProps {
+  item: IQueryItem;
+  comments: ICommentItem[];
+}
