+++ app/dashboard/_components/ItemCard/ItemCard.tsx
@@ -18,15 +18,15 @@ export const ItemCard = ({ item }: { item : IQueryItem | null }) => {
   if (!item) return null;
 
   return (
-    <Card className={`${styles.container}`}>
-      <CardHeader className={`${styles.header}`}>
+    <Card className={`${styles.itemContainer}`}>
+      <CardHeader className={`${styles.itemHeader}`}>
         <Status status={item.status} />
         <FormatedTime time={item.createdAt} />
       </CardHeader>
-      <CardDescription className={`${styles.description}`}>
+      <CardDescription className={`${styles.itemDescription}`}>
         {truncate(item.additionalInfo, 150)}
       </CardDescription>
-      <CardFooter className={`${styles.footer}`}>
+      <CardFooter className={`${styles.itemFooter}`}>
         <Link 
             className={cn(buttonVariants({ variant: 'primary', size: 'sm' }))}
             href={`/dashboard/${item.id}`}
