+++ app/dashboard/_components/ItemCard/styles.module.css
@@ -1,4 +1,4 @@
-.container { 
+.itemContainer { 
   @apply bg-neutral 
     flex 
     flex-col 
@@ -6,18 +6,18 @@
     items-stretch;
 }
 
-.header {
+.itemHeader {
   @apply flex-1 
     flex-row 
     justify-between 
     items-center;
 }
 
-.description {
+.itemDescription {
   @apply flex-1;
 }
 
-.footer {
+.itemFooter {
   @apply flex-1 
     flex-col 
     justify-center
