+++ app/dashboard/_components/QuickView/types.ts
@@ -1,0 +1,8 @@
+// TYPES
+import { IQueryItem } from "@/types/db"
+
+export type QuickViewDialogProps = {
+  item: IQueryItem
+  isOpen: boolean
+  toggleDialog: (option: boolean) => void
+}
