+++ app/signin/_components/SignInForm/types.ts
@@ -1,0 +1,4 @@
+export interface SignInFormProps {
+  email: string;
+  password: string;
+}
